requirements.txt
Flask==0.10.1
Jinja2==2.7.2
MarkupSafe==0.23
Pillow==2.4.0
Werkzeug==0.9.4
itsdangerous==0.24
